Code/requirements.txt
dashscope>=1.14.0
openai>=1.0.0
langchain>=1.0.0
langchain-classic>=1.0.0
langchain-core>=1.0.0
langchain-community>=0.4.0
pyautogui>=0.9.54
pywinauto>=0.6.8
pynput>=1.7.6
mss>=9.0.1
opencv-python>=4.9.0
Pillow>=10.0.0
pytesseract>=0.3.10
pyperclip>=1.8.2
tenacity>=8.2.3
gradio>=6.0.0
pywebview>=4.4.0
chromadb>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
pytest>=7.4.0
pytest-cov>=4.1.0
hypothesis>=6.92.0
jsonschema>=4.20.0
numpy>=1.24.0
